fix: Add Content-Type to /metrics response
When prometheus updates to v3, it is becoming stricter with the
Content-Type of the response.

https://prometheus.io/docs/prometheus/3.0/migration/#scrape-protocols

```
Prometheus v3 is more strict concerning the Content-Type header received when scraping.
Prometheus v2 would default to the standard Prometheus text protocol if the target being
scraped did not specify a Content-Type header or if the header was unparsable or unrecognised.
This could lead to incorrect data being parsed in the scrape. Prometheus v3 will now fail the scrape in such cases.
```

The Content-Type is updated in dependency. So all we have to do is
update the version.

From the metrics/metrics-exporter-prometheus CHANGELOG.md:

```

- Update the `Content-Type` response header to `text/plain`, matching the Exposition format
  specification. ([#496](https://github.com/metrics-rs/metrics/pull/496))
```